+++ android/capacitor/src/main/java/com/getcapacitor/BridgeFragment.java
@@ -62,8 +62,7 @@ public class BridgeFragment extends Fragment {
             startDir = getArguments().getString(ARG_START_DIR);
         }
 
-        bridge =
-            BridgeBuilder.newInstance().setActivity(getActivity()).setInstanceState(savedInstanceState).setPlugins(initialPlugins).build();
+        bridge = new Bridge.Builder().setActivity(getActivity()).setInstanceState(savedInstanceState).setPlugins(initialPlugins).build();
 
         if (startDir != null) {
             bridge.setServerAssetPath(startDir);
